retrieval/requirements.txt
# CVE Retrieval Microservice Requirements
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pymilvus==2.3.4
google-generativeai==0.3.2
python-multipart==0.0.6
requests==2.31.0
numpy==1.25.2
python-dotenv==1.0.0

# Codebase Indexing Requirements
faiss-cpu==1.7.4
flask==3.0.0